requirements.txt
pillow
sqlalchemy
aiogram==3.0.0b7
geopy
redis
timezonefinder
pytz
pyswisseph
apscheduler
tomli_w
yookassa
ephem
pyyaml
python-box
astropy
